+++ examples/example-status/src/main/java/de/otto/edison/example/web/ExampleController.java
@@ -2,6 +2,7 @@ package de.otto.edison.example.web;
 
 import com.codahale.metrics.annotation.Timed;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;