+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * **[edison-mongo]** Fixed display of mongo password in /internal/env
 
 **New Features:**
+* **[edison-core]** Added a UI to configure the log levels of the service. The UI is added to the right nav bar and is
+available under `/internal/loggers`
 * **[edison-core]** Allow LDAP authentication for user-defined paths.
 * **[edison-mongo]** Add @ConditionalOnMissingBean to MongoTogglzRepository
 * **[edison-mongo]** Make ID and ETAG constants public to be able to access them from outside